Remove unused implementations and flatten SimulationAgent
- add argument for the resource manager
- remove unused interface "Response"
